requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.10
Django==5.1.2
django-bootstrap-datepicker==1.5.0
django-bootstrap-datepicker-plus==5.0.5
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-pandas==0.6.7
django-phonenumber-field==8.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
drf-yasg==1.21.8
et-xmlfile==1.1.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
kombu==5.4.2
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==10.4.0
prompt_toolkit==3.0.48
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
reportlab==4.2.5
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.16



